# backend/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ProjectResultDto": {
				"properties": {
					"metric": {
						"type": "string"
					},
					"label": {
						"type": "string"
					}
				},
				"required": [
					"metric",
					"label"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProjectLinksDto": {
				"properties": {
					"live": {
						"type": "string"
					},
					"github": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ProjectDownloadsDto": {
				"properties": {
					"ios": {
						"type": "string"
					},
					"android": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ProjectClientDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"logo": {
						"type": "string"
					},
					"industry": {
						"type": "string"
					},
					"location": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"logo",
					"industry",
					"location"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProjectResponseDto": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"fullDescription": {
						"type": "string"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"features": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"results": {
						"items": {
							"$ref": "#/components/schemas/ProjectResultDto"
						},
						"type": "array"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"team": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"duration": {
						"type": "string"
					},
					"externalLinks": {
						"$ref": "#/components/schemas/ProjectLinksDto"
					},
					"downloadLinks": {
						"$ref": "#/components/schemas/ProjectDownloadsDto"
					},
					"client": {
						"$ref": "#/components/schemas/ProjectClientDto"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"title",
					"description",
					"category",
					"date",
					"images",
					"features",
					"results",
					"tags",
					"team",
					"externalLinks",
					"downloadLinks",
					"client",
					"id",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProjectCreateDto": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"fullDescription": {
						"type": "string"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"features": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"results": {
						"items": {
							"$ref": "#/components/schemas/ProjectResultDto"
						},
						"type": "array"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"team": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"duration": {
						"type": "string"
					},
					"externalLinks": {
						"$ref": "#/components/schemas/ProjectLinksDto"
					},
					"downloadLinks": {
						"$ref": "#/components/schemas/ProjectDownloadsDto"
					},
					"client": {
						"$ref": "#/components/schemas/ProjectClientDto"
					}
				},
				"required": [
					"title",
					"description",
					"category",
					"date",
					"images",
					"features",
					"results",
					"tags",
					"team",
					"externalLinks",
					"downloadLinks",
					"client"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_ProjectCreateDto_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"fullDescription": {
						"type": "string"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"features": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"results": {
						"items": {
							"$ref": "#/components/schemas/ProjectResultDto"
						},
						"type": "array"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"team": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"duration": {
						"type": "string"
					},
					"externalLinks": {
						"$ref": "#/components/schemas/ProjectLinksDto"
					},
					"downloadLinks": {
						"$ref": "#/components/schemas/ProjectDownloadsDto"
					},
					"client": {
						"$ref": "#/components/schemas/ProjectClientDto"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/projects": {
			"get": {
				"operationId": "GetAllProjects",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProjectResponseDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"project": {
											"$ref": "#/components/schemas/ProjectResponseDto"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"project",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProjectCreateDto"
							}
						}
					}
				}
			}
		},
		"/projects/{id}": {
			"get": {
				"operationId": "GetProjectById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ProjectResponseDto"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"project": {
											"$ref": "#/components/schemas/ProjectResponseDto"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ProjectCreateDto_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/investments": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"description": {
												"type": "string"
											},
											"title": {
												"type": "string"
											},
											"email": {
												"type": "string"
											},
											"createdAt": {
												"type": "string",
												"format": "date-time"
											},
											"id": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"description",
											"title",
											"email",
											"createdAt",
											"id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Investments"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"description": {
											"type": "string"
										},
										"title": {
											"type": "string"
										},
										"email": {
											"type": "string"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"description",
										"title",
										"email",
										"createdAt",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Investments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"image": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"title": {
										"type": "string"
									}
								},
								"required": [
									"description",
									"title"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/investments/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"description": {
											"type": "string"
										},
										"title": {
											"type": "string"
										},
										"email": {
											"type": "string"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"description",
										"title",
										"email",
										"createdAt",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Investments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"description": {
											"type": "string"
										},
										"title": {
											"type": "string"
										},
										"email": {
											"type": "string"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"description",
										"title",
										"email",
										"createdAt",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Investments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"image": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"title": {
										"type": "string"
									}
								},
								"required": [
									"description",
									"title"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Investments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/investments/inquiry": {
			"post": {
				"operationId": "CreateInquiry",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"investment": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"investment": {
													"properties": {
														"description": {
															"type": "string"
														},
														"title": {
															"type": "string"
														},
														"email": {
															"type": "string"
														},
														"createdAt": {
															"type": "string",
															"format": "date-time"
														},
														"id": {
															"type": "number",
															"format": "double"
														}
													},
													"required": [
														"description",
														"title",
														"email",
														"createdAt",
														"id"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"investment",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Investments"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"message": {
										"type": "string"
									},
									"investmentNature": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"investmentNature",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}